requirements-dev.txt
# torch # check if really needed
# cv2 # in Jetson opencv is tricky 
numpy 
pillow
matplotlib

onnxruntime 
onnxruntime-gpu


# for loading youtube video
pytubefix
yt_dlp